refactor/change: change source of detailed PR information
Instead of calling the github API each time, yielding a one second delay per
PR query (for the sake of throttling), query the separate repository
https://github.com/jcommelin/gh-mathlib-metadata instead.
That repository has the data downloaded already (updating automatically) ---
effectively providing a cache for these API calls. Additionally, the files
in that repo contain more comprehensive information about each PR:
this change in data format is necessary anyway to roll out #25 and #35.

Particular changes
- remove any reference to print_info.json from dashboard.sh and the script
- instead, download all relevant detail files from the repository
at the beginning of each script run.
(This could be made lazier, or configurable, in the future: this is a good start.)
- inline _print_dashboard, and remove the last argument from it.
- update the parsing of PR information, as the new data has slightly
different format